rename: 패키지명을 entity에서 domain으로 변경
