+++ src/main/java/com/peeerr/climbing/dto/response/CategoryResponse.java
@@ -1,6 +1,6 @@
 package com.peeerr.climbing.dto.response;
 
-import com.peeerr.climbing.entity.Category;
+import com.peeerr.climbing.domain.Category;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
