+++ src/main/java/com/peeerr/climbing/dto/response/PostDetailResponse.java
@@ -1,8 +1,8 @@
 package com.peeerr.climbing.dto.response;
 
-import com.peeerr.climbing.entity.Comment;
-import com.peeerr.climbing.entity.File;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Comment;
+import com.peeerr.climbing.domain.File;
+import com.peeerr.climbing.domain.Post;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
