+++ src/main/java/com/peeerr/climbing/repository/CustomPostRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.peeerr.climbing.repository;
 import com.peeerr.climbing.dto.request.PostSearchCondition;
 import com.peeerr.climbing.dto.response.PostResponse;
 import com.peeerr.climbing.dto.response.QPostResponse;
-import com.peeerr.climbing.entity.Post;
-import com.peeerr.climbing.entity.QPost;
+import com.peeerr.climbing.domain.Post;
+import com.peeerr.climbing.domain.QPost;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.util.StringUtils;
 import com.querydsl.jpa.impl.JPAQuery;
@@ -18,9 +18,9 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-import static com.peeerr.climbing.entity.QCategory.category;
-import static com.peeerr.climbing.entity.QMember.member;
-import static com.peeerr.climbing.entity.QPost.post;
+import static com.peeerr.climbing.domain.QCategory.category;
+import static com.peeerr.climbing.domain.QMember.member;
+import static com.peeerr.climbing.domain.QPost.post;
 
 public class CustomPostRepositoryImpl implements CustomPostRepository {
