+++ src/main/java/com/peeerr/climbing/repository/FileRepository.java
@@ -1,6 +1,6 @@
 package com.peeerr.climbing.repository;
 
-import com.peeerr.climbing.entity.File;
+import com.peeerr.climbing.domain.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FileRepository extends JpaRepository<File, Long> {
