+++ src/main/java/com/peeerr/climbing/repository/LikeRepository.java
@@ -1,7 +1,7 @@
 package com.peeerr.climbing.repository;
 
-import com.peeerr.climbing.entity.Like;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Like;
+import com.peeerr.climbing.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
