+++ src/main/java/com/peeerr/climbing/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.peeerr.climbing.repository;
 
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long>, CustomPostRepository {
