+++ src/main/java/com/peeerr/climbing/security/MemberUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.peeerr.climbing.security;
 
 import com.peeerr.climbing.constant.ErrorMessage;
-import com.peeerr.climbing.entity.Member;
+import com.peeerr.climbing.domain.Member;
 import com.peeerr.climbing.repository.MemberRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
