+++ src/main/java/com/peeerr/climbing/service/CommentService.java
@@ -2,9 +2,9 @@ package com.peeerr.climbing.service;
 
 import com.peeerr.climbing.dto.request.CommentCreateRequest;
 import com.peeerr.climbing.dto.request.CommentEditRequest;
-import com.peeerr.climbing.entity.Comment;
-import com.peeerr.climbing.entity.Member;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Comment;
+import com.peeerr.climbing.domain.Member;
+import com.peeerr.climbing.domain.Post;
 import com.peeerr.climbing.exception.AccessDeniedException;
 import com.peeerr.climbing.exception.notfound.CommentNotFoundException;
 import com.peeerr.climbing.exception.notfound.PostNotFoundException;
