+++ src/test/java/com/peeerr/climbing/integration/docs/CommentDocTest.java
@@ -3,15 +3,15 @@ package com.peeerr.climbing.integration.docs;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.peeerr.climbing.constant.ErrorMessage;
 import com.peeerr.climbing.security.MemberPrincipal;
-import com.peeerr.climbing.entity.Category;
+import com.peeerr.climbing.domain.Category;
 import com.peeerr.climbing.repository.CategoryRepository;
-import com.peeerr.climbing.entity.Comment;
+import com.peeerr.climbing.domain.Comment;
 import com.peeerr.climbing.repository.CommentRepository;
 import com.peeerr.climbing.repository.FileRepository;
 import com.peeerr.climbing.repository.LikeRepository;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Post;
 import com.peeerr.climbing.repository.PostRepository;
-import com.peeerr.climbing.entity.Member;
+import com.peeerr.climbing.domain.Member;
 import com.peeerr.climbing.repository.MemberRepository;
 import com.peeerr.climbing.dto.request.CommentCreateRequest;
 import com.peeerr.climbing.dto.request.CommentEditRequest;
