+++ src/test/java/com/peeerr/climbing/integration/docs/LikeDocTest.java
@@ -1,15 +1,15 @@
 package com.peeerr.climbing.integration.docs;
 
 import com.peeerr.climbing.security.MemberPrincipal;
-import com.peeerr.climbing.entity.Category;
+import com.peeerr.climbing.domain.Category;
 import com.peeerr.climbing.repository.CategoryRepository;
 import com.peeerr.climbing.repository.CommentRepository;
 import com.peeerr.climbing.repository.FileRepository;
-import com.peeerr.climbing.entity.Like;
+import com.peeerr.climbing.domain.Like;
 import com.peeerr.climbing.repository.LikeRepository;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Post;
 import com.peeerr.climbing.repository.PostRepository;
-import com.peeerr.climbing.entity.Member;
+import com.peeerr.climbing.domain.Member;
 import com.peeerr.climbing.repository.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
