+++ src/test/java/com/peeerr/climbing/integration/docs/PostDocTest.java
@@ -24,9 +24,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.peeerr.climbing.constant.ErrorMessage;
 import com.peeerr.climbing.dto.request.PostCreateRequest;
 import com.peeerr.climbing.dto.request.PostEditRequest;
-import com.peeerr.climbing.entity.Category;
-import com.peeerr.climbing.entity.Member;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Category;
+import com.peeerr.climbing.domain.Member;
+import com.peeerr.climbing.domain.Post;
 import com.peeerr.climbing.repository.CategoryRepository;
 import com.peeerr.climbing.repository.CommentRepository;
 import com.peeerr.climbing.repository.FileRepository;
