+++ src/test/java/com/peeerr/climbing/service/LikeServiceTest.java
@@ -1,10 +1,10 @@
 package com.peeerr.climbing.service;
 
-import com.peeerr.climbing.entity.Like;
+import com.peeerr.climbing.domain.Like;
 import com.peeerr.climbing.repository.LikeRepository;
-import com.peeerr.climbing.entity.Post;
+import com.peeerr.climbing.domain.Post;
 import com.peeerr.climbing.repository.PostRepository;
-import com.peeerr.climbing.entity.Member;
+import com.peeerr.climbing.domain.Member;
 import com.peeerr.climbing.repository.MemberRepository;
 import com.peeerr.climbing.exception.already.AlreadyExistsLikeException;
 import com.peeerr.climbing.exception.notfound.CategoryNotFoundException;
